// src/models/mst_year.ts
class mst_year {
    id?: number = 0;
    year_code:                   string = ' '; 
    year:                        string = ' ';
    date_start:                  Date = new Date;
    date_end:                    Date = new Date;
    is_closed:                   boolean = true;
    created_by:                  number = 0;
    created_date:                Date = new Date;
    update_by:                   number = 0;
    update_date:                 Date = new Date;
    is_locked:                   boolean = true;

  }
  
  export default mst_year;
  